Refactored to add optional collision checking.
This refactors collision checking into an optional parameter.  Since
the blender already has a `max_deviation` flag that constrains how far
the blends can differ from the trajectory, it is more reasonable to
use this parameter to prevent collisions than to recheck the entire
trajectory (which is quite slow using CheckPathAllConstraints).